Finish writing the makefile
I finished writing the makefile. It will run the jekyll build command and
deploy it to the server. Updated the _config.yml for the Unix file system
with a case-sensitive file system, which is exactly how it is supposed to
be not some stupid case-insensitive crap like windows. Fix the markdown,
you do NOT indent your lists with a space when writing them, that can break
some markdown parsers. Added a blankline to the gitignore to make it easy
to add to the gitignore as things become apparent about the build, and it
is also the prefered method in my personal style guide.

* Finished writing the makefile